+++ src/LexerParser.h
@@ -1,7 +1,7 @@
 #pragma once
 #ifndef LEXERPARSER_H
 
-#include <stdint.h> // uint8_t
+#include <stdint.h> // uint16_t
 #include <cstdlib> // atoi()
 #include <cctype> // isdigit()
 
@@ -13,12 +13,12 @@ public:
 
 private:
     task todo{ undefined };    // what to do
-    uint8_t num{ 0 };          // number to add or set
+    uint16_t num{ 0 };          // number to add or set
 
 public:
 	void operator()(int argc, char** argv);                  // evaluate task and num
 	const task getTask() const noexcept { return todo; }     // returns task
-    const uint8_t getNum() const noexcept { return num; }    // returns num
+    const uint16_t getNum() const noexcept { return num; }    // returns num
 };
 
 #define LEXERPARSER_H
